backend/requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3
gunicorn==23.0.0
psycopg2-binary==2.9.9
groq==0.13.0
jsonschema==4.23.0
python-dotenv==1.0.0
openpyxl==3.1.2
PyJWT>=2.8.0
Flask-Limiter>=3.5.0
